0.4.1: align reference-doc surface with canonical public API
Documentation discoverability fixes only — no API changes for end users.

The reference docs at reference.langchain.com/python/langchain-parallel
were rendering pages under legacy class names (`ChatParallelWeb`,
`ParallelWebSearchTool`) and at private module paths (`_types`,
`_client`). The user-facing integration docs already promote the
canonical names (`ChatParallel`, `ParallelSearchTool`), so the two
sources drifted. This release reorganizes the source so the
auto-generated reference picks up the canonical names.

Changes:

- `ChatParallel` is now the actual class definition; `ChatParallelWeb`
  is the back-compat alias (`ChatParallelWeb = ChatParallel`). Same
  flip for `ParallelSearchTool` / `ParallelWebSearchTool`. Both names
  continue to import and `isinstance(x, ChatParallelWeb)` continues
  to work — they remain the same class object.
- `langchain_parallel._types` → `langchain_parallel.types`. The
  settings classes (`ExcerptSettings`, `FetchPolicy`,
  `FullContentSettings`, `SourcePolicy`) now live at a public module
  path. `_types` remains as a back-compat re-export shim.
- `langchain_parallel._client` declares `__all__: list[str] = []` and a
  "private module" docstring so doc tooling treats it as internal.

All 87 unit tests pass. ruff + mypy clean.
